# server/openapi.yaml
openapi: 3.0.3
info:
  title: Mystická Hvězda API
  version: 1.0.0
  description: |
    REST API for the Mystická Hvězda astrology platform.
    Provides horoscopes, tarot, numerology, oracle readings, and subscription management.

    **Authentication**: JWT token stored in `auth_token` HttpOnly cookie (set on login).
    **CSRF**: All non-GET requests to `/api/*` require an `X-CSRF-Token` header.
    Get a token from `GET /api/csrf-token` (15-minute TTL).
servers:
  - url: /api
    description: API base path (relative)
  - url: https://mystickahvezda.cz/api
    description: Production
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: HttpOnly JWT cookie set on successful login
  schemas:
...
    AiResponse:
      type: object
      properties:
        success:
          type: boolean
        response:
          type: string
    HoroscopeResponse:
      type: object
      properties:
        success:
          type: boolean
        response:
          type: string
          description: JSON string with prediction, affirmation, luckyNumbers
        period:
          type: string
          example: "Denní inspirace"
        cached:
          type: boolean
...
  parameters:
    XCsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema:
        type: string
      description: CSRF token from GET /api/csrf-token (15-minute TTL)
paths:
...
  /auth/reset-password:
    post:
      tags: [auth]
      summary: Reset password with token
      description: Updates password using Supabase access token from reset email.
      parameters:
        - $ref: '#/components/parameters/XCsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                password:
                  type: string
      responses:
        "200":
          description: Password updated
        "400":
          description: Invalid or expired reset token
...
  /horoscope:
    post:
      tags: [horoscope]
      summary: Get horoscope
      description: |
        Returns daily, weekly, or monthly horoscope.
        Weekly/monthly are premium-only (free users get 402).
        Results cached in DB by sign+period+lang.
      parameters:
        - $ref: '#/components/parameters/XCsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [sign]
              properties:
                sign:
                  type: string
                  enum: [Beran, Býk, Blíženci, Rak, Lev, Panna, Váhy, Štír, Střelec, Kozoroh, Vodnář, Ryby]
                period:
                  type: string
                  enum: [daily, weekly, monthly]
                  default: daily
                context:
                  type: array
                  items:
                    type: string
                  maxItems: 5
                lang:
                  type: string
                  enum: [cs, sk, pl]
                  default: cs
      responses:
        "200":
          description: Horoscope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HoroscopeResponse'
        "400":
          description: Invalid sign or period
        "402":
          description: Weekly/monthly requires premium

  # ─── Numerology ─────────────────────────────────────────────────────────────
  /numerology:
    post:
      tags: [numerology]
      summary: Numerology reading (premium)
      description: |
        Generates a numerology reading from birth date and name.
        Requires authentication and active premium subscription.
        Results cached by input hash.
      security:
        - cookieAuth: []
      parameters:
        - $ref: '#/components/parameters/XCsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, birthDate]
              properties:
                name:
                  type: string
                birthDate:
                  type: string
                  format: date
                birthTime:
                  type: string
                lifePath:
                  type: integer
                destiny:
                  type: integer
                soul:
                  type: integer
                personality:
                  type: integer
      responses:
        "200":
          description: Numerology reading
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiResponse'
        "401":
          description: Not authenticated
        "403":
          description: Premium required

  # ─── Payment ────────────────────────────────────────────────────────────────
...